+++ ChangeLog.md
@@ -5,12 +5,13 @@ This document is created to record all changes.
 ## [Unrelease]
 ### Added
 - Add the header file `includes/ObserverPattern.h` to define the class `myAbstractObserver` for observer pattern.
-- Define the class `mySubjectInterface` for observer pattern.
-- Implement all member methods of the class `mySubjectInterface`.
-- Change the method `mySubjectInterface::notifyData()` from an abstract method to a instance method, and then implement it.
+- Define the class `mySubject` for observer pattern.
+- Implement all member methods of the class `mySubject`.
+- Change the method `mySubject::notifyData()` from an abstract method to a instance method, and then implement it.
 
 ### Changed
 - Rename the class from `myAbstractObserver` to `myObserverInterface`.
+- Rename the class from `mySubjectInterface` to `mySubject` because this class is not an abstract class.
 
 ### Fixed